+++ csvtk/cmd/root.go
@@ -64,6 +64,9 @@ Attention:
          extraneous or missing " in quoted-field
      Please use "csvtk fix-quotes" to fix it, and use "csvtk del-quotes" to reset to the
      original format as needed.
+  9. csvtk writes gzip files very fast, much faster than the multi-threaded pigz,
+     therefore there's no need to pipe the result to gzip/pigz.
+     csvtk also supports reading and writing xz (.xz), zstd (.zst) and Bzip2 (.bz2) formats.
 
 Environment variables for frequently used global flags:
 
@@ -71,7 +74,7 @@ Environment variables for frequently used global flags:
   - "CSVTK_H" for flag "-H/--no-header-row"
   - "CSVTK_QUIET" for flag "--quiet"
 
-You can also create a soft link named "tsvtk" for "csvtk", 
+You can also create a soft link named "tsvtk" for "csvtk",
 which sets "-t/--tabs" by default.
 
 `, VERSION),